requirements.txt
groq
python-dotenv
gradio==4.12.0

